Extrait les fonctions dans un fichier dédié
